Move Version to Infrastructure
