+++ src/Command/Release/MakeCommand.php
@@ -5,14 +5,13 @@ namespace Yiisoft\YiiDevTool\Command\Release;
 use GitWrapper\GitWorkingCopy;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 use Yiisoft\YiiDevTool\Component\Changelog;
 use Yiisoft\YiiDevTool\Component\Console\PackageCommand;
 use Yiisoft\YiiDevTool\Component\Package\Package;
-use Yiisoft\YiiDevTool\Component\Version;
+use Yiisoft\YiiDevTool\Infrastructure\Version;
 
 class MakeCommand extends PackageCommand
 {
